fix rfregressor

+++ fpl_reco/getrecs.py
@@ -122,7 +122,7 @@ def main():
 	connection = connectDB()
 	tuples = runCommand(connection)
 	print(pd.DataFrame([item[3:18] for item in tuples]))
-	rfregressor = RandomForest(pd.DataFrame([item[3:18] for item in tuples]), np.array([item[-1] for item in tuples]), 15, 'sqrt', 592)
+	rfregressor = RandomForest(pd.DataFrame([item[3:18] for item in tuples], dtype=float), np.array([item[-1] for item in tuples]), 15, 'sqrt', 592)
 	pred_scores = []
 	for i in tuples:
 		pred_scores += [rfregressor.predict(np.array(i[3:18]))]
